pyproject.toml
[tool.poetry]
name = "pyhealthcheck"
version = "0.1.0-alpha"
description = "FastAPI project generated using minimal-fastapi-postgres-template."
authors = ["admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.68.1"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.26"}
uvicorn = "^0.15.0"
python-dotenv = "^0.19.0"
requests = "^2.26.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
pydantic = {version = "^1.8.2", extras = ["email"]}
alembic = "^1.7.3"
python-multipart = "^0.0.5"
asyncpg = "^0.24.0"
aioping = "^0.3.1"
uvloop = "^0.16.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.9b0", python = ">=3.6.2,<4.0.0"}
autoflake = "^1.4"
flake8 = "^3.9.2"
isort = "^5.9.3"
coverage = "^5.5"
pytest = "^6.2.5"
mypy = "^0.910"
pytest-asyncio = "^0.16.0"
httpx = "^0.20.0"
sqlalchemy-stubs = "^0.4"
sqlalchemy2-stubs = "^0.0.2-alpha.19"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v"
testpaths = [
    "app/tests",
]

[tool.isort]
profile = "black"

# [tool.pyright]
# reportUnusedImport = "warning"
# reportMissingTypeStubs = "none"
# reportUnknownVariableType = "warning"
# reportUntypedFunctionDecorator = "warning"
# reportIncompatibleVariableOverride = "none"
# reportGeneralTypeIssues = "warning"
# reportUnknownMemberType = "none"